moving things around; fixed a bug where the inspection happened before image_name abbreviations were expanded
